refactor(dispatch): separate parsing and projections
