+++ src/app/dispatch/orchestration.rs
@@ -1,6 +1,7 @@
 //! Orchestration JSON API handlers.
 
 use super::*;
+use super::{params::*, projection::*};
 
 impl App {
     // ── worktrees (docs/18 WT-3) ──
@@ -646,5 +647,388 @@ impl App {
             "type": "lease_list",
             "leases": serde_json::to_value(&self.orch.leases).unwrap_or(Value::Null),
         }))
+    }
+}
+
+/// An orchestration `Reject` → the API `(code, message)` error shape.
+pub(in crate::app::dispatch) fn orch_err(r: crate::orch::Reject) -> (String, String) {
+    (r.code.to_string(), r.message)
+}
+
+pub(in crate::app::dispatch) fn automation_err(r: crate::automation::Reject) -> (String, String) {
+    (r.code.to_string(), r.message)
+}
+
+pub(in crate::app::dispatch) fn automation_trigger(
+    value: Option<&Value>,
+) -> Result<crate::automation::Trigger, (String, String)> {
+    let value = value.ok_or_else(|| {
+        (
+            "invalid_request".to_string(),
+            "trigger is required".to_string(),
+        )
+    })?;
+    let kind = value.get("kind").and_then(Value::as_str).ok_or_else(|| {
+        (
+            "invalid_schedule".to_string(),
+            "trigger.kind is required".to_string(),
+        )
+    })?;
+    reject_api_fields(
+        value,
+        match kind {
+            "once" => &["kind", "at_utc"],
+            "interval" => &["kind", "every_seconds", "anchor_utc"],
+            "daily" => &["kind", "timezone", "second_of_day"],
+            "weekly" => &["kind", "timezone", "weekdays", "second_of_day"],
+            _ => {
+                return Err((
+                    "invalid_schedule".to_string(),
+                    format!("unknown trigger kind: {kind}"),
+                ))
+            }
+        },
+    )?;
+    serde_json::from_value(value.clone()).map_err(|error| {
+        (
+            "invalid_schedule".to_string(),
+            format!("invalid trigger: {error}"),
+        )
+    })
+}
+
+pub(in crate::app::dispatch) fn automation_input(
+    p: &Value,
+) -> Result<crate::automation::CreateAutomation, (String, String)> {
+    let task = p.get("task").and_then(Value::as_object).ok_or_else(|| {
+        (
+            "invalid_request".to_string(),
+            "task object is required".to_string(),
+        )
+    })?;
+    reject_api_fields(
+        p.get("task").expect("task was just validated"),
+        &[
+            "title",
+            "prompt",
+            "agent_id",
+            "workspace_id",
+            "mode",
+            "access",
+            "paths",
+            "gate",
+        ],
+    )?;
+    if let Some(policy) = p.get("policy") {
+        reject_api_fields(policy, &["misfire", "overlap", "misfire_grace_seconds"])?;
+    }
+    let policy = p
+        .get("policy")
+        .cloned()
+        .map(serde_json::from_value)
+        .transpose()
+        .map_err(|error| ("invalid_policy".to_string(), error.to_string()))?
+        .unwrap_or_default();
+    let mode = match task.get("mode") {
+        None => crate::orch::TaskWorkerMode::Worktree,
+        Some(Value::String(mode)) => crate::orch::TaskWorkerMode::parse(mode).ok_or_else(|| {
+            (
+                "invalid_request".to_string(),
+                "task.mode must be worktree or workspace".to_string(),
+            )
+        })?,
+        Some(_) => {
+            return Err((
+                "invalid_request".to_string(),
+                "task.mode must be a string".to_string(),
+            ))
+        }
+    };
+    let access = match task.get("access") {
+        None => crate::automation::AutomationAccess::default(),
+        Some(Value::String(access)) => crate::automation::AutomationAccess::parse(access)
+            .ok_or_else(|| {
+                (
+                    "invalid_request".to_string(),
+                    "task.access must be read_only, workspace, or full_access".to_string(),
+                )
+            })?,
+        Some(_) => {
+            return Err((
+                "invalid_request".to_string(),
+                "task.access must be a string".to_string(),
+            ))
+        }
+    };
+    let enabled = match p.get("enabled") {
+        None => true,
+        Some(Value::Bool(enabled)) => *enabled,
+        Some(_) => {
+            return Err((
+                "invalid_request".to_string(),
+                "enabled must be a boolean".to_string(),
+            ))
+        }
+    };
+    let paths = match task.get("paths") {
+        None => Vec::new(),
+        Some(Value::Array(values)) => values
+            .iter()
+            .map(|value| {
+                value.as_str().map(str::to_string).ok_or_else(|| {
+                    (
+                        "invalid_request".to_string(),
+                        "task.paths must contain only strings".to_string(),
+                    )
+                })
+            })
+            .collect::<Result<Vec<_>, _>>()?,
+        Some(_) => {
+            return Err((
+                "invalid_request".to_string(),
+                "task.paths must be an array".to_string(),
+            ))
+        }
+    };
+    let gate = match task.get("gate") {
+        None | Some(Value::Null) => None,
+        Some(Value::String(gate)) => Some(gate.trim().to_string()).filter(|gate| !gate.is_empty()),
+        Some(_) => {
+            return Err((
+                "invalid_request".to_string(),
+                "task.gate must be a string or null".to_string(),
+            ))
+        }
+    };
+    Ok(crate::automation::CreateAutomation {
+        name: req_str(p, "name")?.to_string(),
+        enabled,
+        trigger: automation_trigger(p.get("trigger"))?,
+        target: automation_target(p.get("target"))?,
+        task: crate::automation::TaskTemplate {
+            title: task
+                .get("title")
+                .and_then(Value::as_str)
+                .ok_or_else(|| {
+                    (
+                        "invalid_request".to_string(),
+                        "task.title is required".to_string(),
+                    )
+                })?
+                .to_string(),
+            prompt: task
+                .get("prompt")
+                .and_then(Value::as_str)
+                .ok_or_else(|| {
+                    (
+                        "invalid_request".to_string(),
+                        "task.prompt is required".to_string(),
+                    )
+                })?
+                .to_string(),
+            agent_id: task
+                .get("agent_id")
+                .and_then(Value::as_str)
+                .ok_or_else(|| {
+                    (
+                        "invalid_request".to_string(),
+                        "task.agent_id is required".to_string(),
+                    )
+                })?
+                .to_string(),
+            workspace_id: task
+                .get("workspace_id")
+                .and_then(Value::as_str)
+                .ok_or_else(|| {
+                    (
+                        "invalid_request".to_string(),
+                        "task.workspace_id is required".to_string(),
+                    )
+                })?
+                .to_string(),
+            mode,
+            access,
+            paths,
+            gate,
+        },
+        policy,
+    })
+}
+
+pub(in crate::app::dispatch) fn automation_target(
+    value: Option<&Value>,
+) -> Result<crate::automation::AutomationTarget, (String, String)> {
+    use crate::automation::{ActiveAgentBusyPolicy, AutomationTarget};
+
+    let Some(value) = value else {
+        return Ok(AutomationTarget::NewWorker);
+    };
+    let object = value.as_object().ok_or_else(|| {
+        (
+            "invalid_target".to_string(),
+            "target must be an object".to_string(),
+        )
+    })?;
+    reject_api_fields(value, &["kind", "pane_id", "terminal_id", "if_busy"])?;
+    match object.get("kind").and_then(Value::as_str) {
+        Some("new_worker") => {
+            if object.len() != 1 {
+                return Err((
+                    "invalid_target".to_string(),
+                    "new_worker target accepts only kind".to_string(),
+                ));
+            }
+            Ok(AutomationTarget::NewWorker)
+        }
+        Some("active_agent") => {
+            let pane_id = object
+                .get("pane_id")
+                .and_then(|value| {
+                    value
+                        .as_str()
+                        .and_then(|value| value.parse::<u32>().ok())
+                        .or_else(|| value.as_u64().and_then(|value| u32::try_from(value).ok()))
+                })
+                .filter(|pane| *pane != 0)
+                .ok_or_else(|| {
+                    (
+                        "invalid_target".to_string(),
+                        "active_agent target requires a non-zero pane_id".to_string(),
+                    )
+                })?;
+            let terminal_id = object
+                .get("terminal_id")
+                .and_then(Value::as_str)
+                .filter(|value| !value.is_empty())
+                .ok_or_else(|| {
+                    (
+                        "invalid_target".to_string(),
+                        "active_agent target requires terminal_id".to_string(),
+                    )
+                })?
+                .to_string();
+            let if_busy = match object.get("if_busy").and_then(Value::as_str) {
+                None | Some("wait") => ActiveAgentBusyPolicy::Wait,
+                Some("skip") => ActiveAgentBusyPolicy::Skip,
+                Some(_) => {
+                    return Err((
+                        "invalid_target".to_string(),
+                        "target.if_busy must be wait or skip".to_string(),
+                    ))
+                }
+            };
+            Ok(AutomationTarget::ActiveAgent {
+                pane_id,
+                terminal_id,
+                if_busy,
+                durable: None,
+            })
+        }
+        _ => Err((
+            "invalid_target".to_string(),
+            "target.kind must be new_worker or active_agent".to_string(),
+        )),
+    }
+}
+
+pub(in crate::app::dispatch) fn validate_automation_target(
+    app: &App,
+    input: &mut crate::automation::CreateAutomation,
+) -> Result<(), (String, String)> {
+    if let crate::automation::AutomationTarget::ActiveAgent { .. } = &input.target {
+        app.prepare_active_agent_target(&mut input.target, &mut input.task)?;
+        return Ok(());
+    }
+
+    let descriptor = crate::agent::registry::find(&input.task.agent_id).ok_or_else(|| {
+        (
+            "unsupported_agent".to_string(),
+            format!(
+                "{} is not a launch-capable built-in agent",
+                input.task.agent_id
+            ),
+        )
+    })?;
+    input.task.agent_id = descriptor.id.to_string();
+    if !descriptor
+        .automation
+        .is_some_and(|operations| operations.supports(input.task.access))
+    {
+        return Err((
+            "unsupported_automation_access".to_string(),
+            format!(
+                "{} does not support {} scheduled access",
+                descriptor.id,
+                input.task.access.label().to_ascii_lowercase()
+            ),
+        ));
+    }
+    if !app
+        .workspaces
+        .iter()
+        .any(|workspace| workspace.id == input.task.workspace_id)
+    {
+        return Err((
+            "workspace_not_found".to_string(),
+            format!("workspace id {} not found", input.task.workspace_id),
+        ));
+    }
+    // Reuse ORCH's title/path/gate validation without mutating the live ledger.
+    let mut probe = crate::orch::OrchState::default();
+    probe
+        .add_task(
+            input.task.title.clone(),
+            input.task.paths.clone(),
+            Vec::new(),
+            input.task.gate.clone(),
+        )
+        .map_err(orch_err)?;
+    Ok(())
+}
+
+pub(in crate::app::dispatch) fn task_worker_mode(
+    p: &Value,
+    existing: Option<crate::orch::TaskWorkerMode>,
+) -> Result<crate::orch::TaskWorkerMode, (String, String)> {
+    match p.get("mode") {
+        None => Ok(existing.unwrap_or(crate::orch::TaskWorkerMode::Worktree)),
+        Some(Value::String(mode)) => crate::orch::TaskWorkerMode::parse(mode).ok_or_else(|| {
+            (
+                "bad_request".to_string(),
+                "mode must be worktree or workspace".to_string(),
+            )
+        }),
+        Some(_) => Err((
+            "bad_request".to_string(),
+            "mode must be worktree or workspace".to_string(),
+        )),
+    }
+}
+
+pub(in crate::app::dispatch) fn optional_task_prompt(
+    p: &Value,
+) -> Result<Option<String>, (String, String)> {
+    match p.get("prompt") {
+        None | Some(Value::Null) => Ok(None),
+        Some(Value::String(prompt)) => Ok(Some(prompt.clone())),
+        Some(_) => Err((
+            "invalid_request".to_string(),
+            "prompt must be a string or null".to_string(),
+        )),
+    }
+}
+
+impl App {
+    /// The pane a task/lease call acts for: the passed `pane`, else the caller's
+    /// `$LUVUS_PANE_ID`. Orchestration is pane-keyed, so this is required.
+    pub(in crate::app::dispatch) fn orch_pane(&self, p: &Value) -> Result<u32, (String, String)> {
+        self.resolve_optional_pane(p)?
+            .map(|id| id.0)
+            .ok_or_else(|| {
+                (
+                    "no_pane".to_string(),
+                    "no pane id — run inside a luvus pane or pass a pane id".to_string(),
+                )
+            })
     }
 }
