+++ src/app/dispatch/runtime.rs
@@ -1,5 +1,6 @@
 //! Event-driven agent detection and runtime scan scheduling.
 
+use super::projection::*;
 use super::*;
 use std::sync::Arc;
 
@@ -873,4 +874,67 @@ impl App {
         }
         repaired_location || changed
     }
+}
+
+pub(in crate::app::dispatch) fn log_agent_identity(id: PaneId, agent: &str, source: &str) {
+    let authority = match source {
+        "integration_report" => crate::logging::Authority::Hook,
+        source if source.contains("process") => crate::logging::Authority::Process,
+        "command_fallback" => crate::logging::Authority::None,
+        _ => crate::logging::Authority::Text,
+    };
+    let mut fields = [crate::logging::Field::IdOmitted(false); 4];
+    fields[0] = crate::logging::Field::PaneId(u64::from(id.0));
+    fields[1] = crate::logging::Field::Authority(authority);
+    let count = if let Some(agent) = crate::logging::SafeId::new(agent) {
+        fields[2] = crate::logging::Field::Agent(agent);
+        3
+    } else {
+        fields[2] = crate::logging::Field::IdOmitted(true);
+        3
+    };
+    crate::logging::event(crate::logging::EventKind::AgentIdentity, &fields[..count]);
+}
+
+pub(in crate::app::dispatch) fn log_agent_state(id: PaneId, agent: &str, from: State, to: State) {
+    fn map(state: State) -> crate::logging::AgentState {
+        match state {
+            State::Blocked => crate::logging::AgentState::Blocked,
+            State::Working => crate::logging::AgentState::Working,
+            State::Done => crate::logging::AgentState::Done,
+            State::Idle | State::Unknown => crate::logging::AgentState::Idle,
+        }
+    }
+
+    let mut fields = [crate::logging::Field::IdOmitted(false); 5];
+    fields[0] = crate::logging::Field::PaneId(u64::from(id.0));
+    fields[1] = crate::logging::Field::FromState(map(from));
+    fields[2] = crate::logging::Field::AgentState(map(to));
+    let count = if let Some(agent) = crate::logging::SafeId::new(agent) {
+        fields[3] = crate::logging::Field::Agent(agent);
+        4
+    } else {
+        fields[3] = crate::logging::Field::IdOmitted(true);
+        4
+    };
+    crate::logging::event(crate::logging::EventKind::AgentState, &fields[..count]);
+}
+
+pub(in crate::app::dispatch) fn log_agent_authority(
+    id: PaneId,
+    agent: &str,
+    outcome: crate::logging::Outcome,
+) {
+    let mut fields = [crate::logging::Field::IdOmitted(false); 5];
+    fields[0] = crate::logging::Field::PaneId(u64::from(id.0));
+    fields[1] = crate::logging::Field::Authority(crate::logging::Authority::Hook);
+    fields[2] = crate::logging::Field::Outcome(outcome);
+    let count = if let Some(agent) = crate::logging::SafeId::new(agent) {
+        fields[3] = crate::logging::Field::Agent(agent);
+        4
+    } else {
+        fields[3] = crate::logging::Field::IdOmitted(true);
+        4
+    };
+    crate::logging::event(crate::logging::EventKind::AgentAuthority, &fields[..count]);
 }
