+++ src/app/dispatch/tests/api.rs
@@ -1,3 +1,4 @@
+use super::params::*;
 use super::*;
 use crate::app::App;
